display single post in view post

+++ viewPost.php
@@ -1,4 +1,14 @@
-<?php require_once 'inc/header.php' ?>
+<?php require_once 'inc/header.php';
+require_once 'config/connection.php';
+
+if (isset($_GET['id'])) {
+  $id = $_GET['id'];
+}
+$query = "SELECT * FROM posts where id = '$id';";
+$result = mysqli_query($connection, $query);
+$fetch_posts = mysqli_fetch_all($result, MYSQLI_ASSOC);
+
+?>
 <link rel="stylesheet" href="assets/css/viewPost.css">
 
 <div class="page-heading products-heading header-text">
@@ -22,28 +32,33 @@
           <h2>Post Details</h2>
         </div>
       </div>
-      <div class="col-md-6">
-        <div class="right-image">
-          <img src="assets/images/feature-image.jpg" alt="Post cover">
-        </div>
-      </div>
-      <div class="col-md-6">
-        <div class="left-content">
-          <div class="post-detail__meta">
-            <span class="post-date"><i class="fa fa-calendar"></i> 2026-05-12</span>
-            <span class="tag-chip">Blog</span>
+      <?php if ($fetch_posts):
+        foreach ($fetch_posts as $post): ?>
+          <div class="col-md-6">
+            <div class="right-image">
+              <img src="assets/images/postImage/<?= $post['image'] ?>" alt="Post cover">
+            </div>
           </div>
-          <h4>Who we are &amp; What we do?</h4>
-          <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Sed voluptate nihil eum consectetur similique? Consectetur, quod, incidunt, harum nisi dolores delectus reprehenderit voluptatem perferendis dicta dolorem non blanditiis ex fugiat. Lorem ipsum dolor sit amet, consectetur adipisicing elit.<br><br>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Et, consequuntur, modi mollitia corporis ipsa voluptate corrupti eum ratione ex ea praesentium quibusdam? Aut, in eum facere corrupti necessitatibus perspiciatis quis.</p>
-          <div class="post-detail__actions">
-            <a href="index.php" class="btn-back">
-              <i class="fa fa-long-arrow-left"></i> Back to posts
-            </a>
-            <a href="editPost.php" class="btn btn-outline-primary">Edit post</a>
-            <a href="deletePost.php" class="btn btn-danger">Delete post</a>
+          <div class="col-md-6">
+            <div class="left-content">
+              <div class="post-detail__meta">
+                <span class="post-date"><i class="fa fa-calendar"></i> <?= $post['created_at']; ?></span>
+                <span class="tag-chip">Blog</span>
+              </div>
+              <h4><?= $post['title']; ?></h4>
+              <p><?= $post['body']; ?></p>
+              <div class="post-detail__actions">
+                <a href="index.php" class="btn-back">
+                  <i class="fa fa-long-arrow-left"></i> Back to posts
+                </a>
+                <a href="editPost.php" class="btn btn-outline-primary">Edit post</a>
+                <a href="deletePost.php" class="btn btn-danger">Delete post</a>
+              </div>
+            </div>
           </div>
-        </div>
-      </div>
+      <?php
+        endforeach;
+      endif; ?>
     </div>
   </div>
 </div>
